requirements.txt
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
pymupdf
pytesseract
Pillow
langgraph
neo4j
chromadb
streamlit
streamlit-agraph
langchain-neo4j
langchain-chroma
python-dotenv
pypdf
fastapi
uvicorn[standard]
python-multipart
ragas
pandas
